CodeRepos/Radar/requirements.txt
absl-py==1.4.0
action-tutorials-py==0.9.4
adversarial-robustness-toolbox==1.13.0
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
altgraph==0.17.4
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
astor==0.8.0
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.2
atari-py==0.2.6
atomicwrites==1.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==6.1.0
blessings==1.7
brotlipy==0.7.0
build==0.10.0
cachetools==4.1.1
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cleverhans==4.0.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.3
colorful==0.5.4
configparser==5.0.1
contextvars==2.4
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cssselect==1.2.0
cssutils==2.7.1
cycler==0.10.0
Cython==0.29.21
decorator==5.1.1
deepdiff==4.3.2
defusedxml==0.7.1
demo-nodes-py==0.9.4
dill==0.3.2
dm-tree==0.1.8
doc2docx==0.2.4
docker-pycreds==0.4.0
docopt==0.6.2
domain-coordinator==0.9.2
dpcpp-cpp-rt==2023.0.0
eagerpy==0.30.0
easydict==1.10
einops==0.8.0
enum34==1.1.10
et-xmlfile==1.1.0
evdev==1.6.1
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.15
fastjsonschema==2.20.0
filelock==3.0.12
flatbuffers==23.1.4
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
foolbox==3.3.3
frozenlist==1.3.3
fsspec==2023.3.0
funcsigs==1.0.2
future==0.16.0
gast==0.5.3
gin==0.1.6
gin-config==0.3.0
gitdb==4.0.10
GitPython==3.1.30
glfw==1.12.0
google==3.0.0
google-api-core==1.22.1
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
gql==0.2.0
graphql-core==1.1
grpcio==1.51.1
gym==0.26.2
-e git+https://github.com/hanruihua/rl_rvo_nav.git@c5c1a6ad2f81c004b830019f6a79d75b6af66799#egg=gym_env&subdirectory=gym_env
gym-notices==0.0.8
h5py==3.7.0
hiredis==1.1.0
idna==2.10
idna-ssl==1.1.0
imageio==2.4.1
immutables==0.14
importlib-metadata==6.6.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
ipdb==0.13.13
ipython==8.12.3
-e git+https://ghproxy.com/https://github.com/hanruihua/intelligent-robot-simulator.git@c74fcd24bf8bb43fe36ff1d1a0b7fcc899851b3d#egg=ir_sim
jedi==0.19.0
Jinja2==3.1.4
joblib==1.2.0
jsonnet==0.16.0
jsonpickle==0.9.6
jsonschema==3.2.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
libclang==15.0.6.1
lightning-utilities==0.8.0
lockfile==0.12.2
lxml==4.9.2
Markdown==3.1.1
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
mkl==2023.0.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mnist==0.2.2
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
mpyq==0.2.5
msgpack==1.0.0
mujoco-py==2.0.2.8
multidict==6.0.4
munch==2.3.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opencv-python==4.7.0.68
openpyxl==3.1.5
opt-einsum==3.3.0
ordered-set==4.0.2
osrf-pycommon==0.1.11
packaging==20.4
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
path==16.6.0
path.py==12.5.0
pathlib==1.0.1
pathlib2==2.3.2
pathtools==0.1.2
pbr==4.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip-tools==6.13.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.14.1
pluggy==0.7.1
portpicker==1.2.0
premailer==3.10.0
probscale==0.2.3
progressbar2==3.53.1
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.6.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.20
pygame==2.1.2
pyglet==1.5.0
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pykalman==0.9.5
pynput==1.7.6
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.16.0
PySC2==3.0.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytest==3.8.2
python-dateutil==2.9.0.post0
python-utils==2.4.0
python-xlib==0.33
pytorch-lightning==2.0.0
pytz==2020.1
PyYAML==6.0
pyzmq==26.0.3
quality-of-service-demo-py==0.9.4
redis==3.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ros2action==0.9.13
ros2bag==0.3.11
ros2cli==0.9.13
ros2component==0.9.13
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
rsa==4.6
s2clientprotocol==4.10.1.75800.0
s2protocol==4.11.4.78285.0
sacred==0.7.2
scikit-learn==1.1.3
scipy==1.10.0
seaborn==0.10.1
sentry-sdk==1.15.0
setproctitle==1.3.2
shortuuid==1.0.1
six==1.15.0
sk-video==1.1.10
smmap==5.0.0
snakeviz==1.0.0
soupsieve==2.0.1
sros2==0.9.5
stack-data==0.6.2
subprocess32==3.5.4
swin-transformer-pytorch==0.4.1
sympy==1.12
tabulate==0.8.7
tbb==2021.8.0
teleop-twist-keyboard==2.3.2
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.8.4
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
tinycss2==1.3.0
tomli==2.0.1
topic-monitor==0.9.4
torch==2.1.0
torchattacks==3.4.0
torchaudio==0.13.1
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchvision==0.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
triton==2.1.0
typing-extensions==3.7.4.3
urllib3==1.23
wandb==0.13.10
watchdog==0.10.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.16.1
whichcraft==0.5.2
wrapt==1.12.1
xmltodict==0.12.0
yagmail==0.15.293
yarg==0.1.9
yarl==1.8.2
zipp==3.1.0
zmq==0.0.0
